Add argument to build static diles to anable building just choosen pages

+++ static_site_generator/sitebuilder/management/commands/build.py
@@ -3,7 +3,7 @@ import shutil
 
 from django.conf import settings
 from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.core.urlresolvers import reverse
 from django.test.client import Client
 
@@ -19,22 +19,38 @@ class Command(BaseCommand):
     help = 'Build static site output.'
     leave_local_alone = True
 
+    def add_argument(self, parser):
+        parser.add_argument('args', nargs='*')
+
     def handle(self, *args, **options):
         """Request pages and build output."""
-        if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
-            shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
-        os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
+        if args:
+            pages = args
+            available = list(get_pages())
+            invalid = []
+            for page in pages:
+                if page not in available:
+                    invalid.append(page)
+            if invalid:
+                msg = 'Invalid pages: {}'.format(', '.join(invalid))
+                raise CommandError(msg)
+        else:
+            pages = get_pages()
+            if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
+                shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
+            os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
         os.makedirs(settings.STATIC_ROOT, exist_ok=True)
         call_command('collectstatic', interactive=False,
-            clear=True, verbosity=0)
+             clear=True, verbosity=0)
         client=Client()
-        for page in get_pages():
+        for page in pages:
             url = reverse('page', kwargs={'slug':page})
             response = client.get(url)
             if page == 'index':
                 output_dir = settings.SITE_OUTPUT_DIRECTORY
             else:
                 output_dir = os.path.join(settings.SITE_OUTPUT_DIRECTORY, page)
-                os.makedirs(output_dir)
+                if not os.path.exists(output_dir):
+                    os.makedirs(output_dir)
             with open(os.path.join(output_dir, 'index.html'), 'wb') as f:
                 f.write(response.content)
